修复了 无法加载https（不受信任的证书）图片bug； 修复了 SLWebViewController无法访问https（不受信任的证书）地址bug；

+++ ug/Classes/六合模板/Controller/六合页面/LHJournalDetailVC.m
@@ -43,7 +43,7 @@
     
     __weakSelf_(__self);
     // 六合图库的收藏按钮
-    if (_gm.gid) {
+    if (_gm.gid.intValue) {
         self.navigationItem.rightBarButtonItem = [STBarButtonItem barButtonItemWithTitle:@"收藏" block:^(UIButton *sender) {
             if (!UGLoginIsAuthorized()) {
                 SANotificationEventPost(UGNotificationShowLoginView, nil);
